Add workaround for DRG jumps; refactor/reorganize
As noted in the code, DRG jumps appear in the abilities table in DSP but
are sent as a WEAPONSKILL_FINISH packet. They would normally break the
addon's display, so just ignore them for now.

Also refactored the code to rely on the libs I've made and commented a
lot of what I wrote.