Unstub debuff bar in tgt
This code is heavily inspired by xipie's myfx addon, which fulfills a
similar need. It has no license, but it's on github so I'm assuming the
author is okay with people using their code like that.

There are probably still bugs, and there's definitely some uncertainty
around whether or not I'm correctly tracking dropped statuses. Currently
we are totally ignoring basic message packets, so we're 100% relying on
the default timers being accurate (they probably aren't though).