Add rcheck
This is a simple tool to handle ready checks for raids. It sends a
message indicating what's happening, and tallies the people that respond
in party chat over the next 20 seconds. At the end, it will give a count
and the names of those not ready yet.

There are a few extra stop-in points every 5 seconds. I could probably
rework the timer code to be smaller and run more often, but I don't care
that much yet.

I also don't know how to have injected packets appear in the sender's
chat log, so for those we're just showing an addon message. This also
lets us silently check in with the user without spamming the alliance.

The ready checker is automatically counted as ready.