Correctly count alliance members
Like I did with `skillchain`, I had the party-checking function bailing
early on the first inactive slow - which is not how that API should be
used.

In an effort to have the addon assume the readiness of the caller, it
skips counting the first party slot and just adds 1 to the counts but
does not exclude them from the party-chat monitoring. This could lead to
double-counting the caller.