Monitor party chat for specific strings in rcheck
It was a poor assumption that if a player says something, then they're
ready. Additionally, our linkshell has a strong culture of communication
and paying attention (or the officers do, anyway). This leads me to
favor more particular chat monitoring - this way, someone who says "afk"
won't be counted as ready; but people who say "ready" will be called out
and can be quickly identified as actually ready.